requirements.txt
# crytography package
pycryptodome==3.15.0
